requirements.txt
streamlit
google-generativeai
langgraph
langchain-community
duckduckgo-search